feat: refactor a little and websockets
