+++ public/src/environments/environment.prod.ts
@@ -1,4 +1,5 @@
 export const environment = {
   production: true,
-  statusUrl: '/status'
+  statusUrl: '/status',
+  webSocketUrl: 'wss://'+window.location.host
 };
