+++ server/package.json
@@ -13,7 +13,8 @@
   "license": "MIT",
   "dependencies": {
     "express": "^4.17.1",
-    "source-map-support": "^0.5.19"
+    "source-map-support": "^0.5.19",
+    "ws": "^7.4.4"
   },
   "devDependencies": {
     "@types/express": "^4.17.11",
